third_party/blink/web_tests/virtual/non-streaming-declarative-shadow-dom-enabled/non-streaming-tests/getinnerhtml.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Element.getInnerHTML() on <article>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<article><template shadowroot=\\"open\\"><slot></slot></template></article>" but got "<article><template shadowrootmode=\\"open\\"><slot></slot></template></article>"
[FAIL] Element.getInnerHTML() on <article>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<article><template shadowroot=\\"closed\\"><slot></slot></template></article>" but got "<article><template shadowrootmode=\\"closed\\"><slot></slot></template></article>"
[FAIL] Element.getInnerHTML() on <article>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<article><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></article>" but got "<article><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></article>"
[FAIL] Element.getInnerHTML() on <article>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<article><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></article>" but got "<article><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></article>"
[FAIL] ShadowRoot.getInnerHTML() on <article>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<article><template shadowroot=\\"open\\"><slot></slot></template></article>" but got "<article><template shadowrootmode=\\"open\\"><slot></slot></template></article>"
[FAIL] ShadowRoot.getInnerHTML() on <article>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<article><template shadowroot=\\"closed\\"><slot></slot></template></article>" but got "<article><template shadowrootmode=\\"closed\\"><slot></slot></template></article>"
[FAIL] ShadowRoot.getInnerHTML() on <article>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<article><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></article>" but got "<article><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></article>"
[FAIL] ShadowRoot.getInnerHTML() on <article>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<article><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></article>" but got "<article><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></article>"
[FAIL] Element.getInnerHTML() on <aside>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<aside><template shadowroot=\\"open\\"><slot></slot></template></aside>" but got "<aside><template shadowrootmode=\\"open\\"><slot></slot></template></aside>"
[FAIL] Element.getInnerHTML() on <aside>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<aside><template shadowroot=\\"closed\\"><slot></slot></template></aside>" but got "<aside><template shadowrootmode=\\"closed\\"><slot></slot></template></aside>"
[FAIL] Element.getInnerHTML() on <aside>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<aside><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></aside>" but got "<aside><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></aside>"
[FAIL] Element.getInnerHTML() on <aside>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<aside><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></aside>" but got "<aside><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></aside>"
[FAIL] ShadowRoot.getInnerHTML() on <aside>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<aside><template shadowroot=\\"open\\"><slot></slot></template></aside>" but got "<aside><template shadowrootmode=\\"open\\"><slot></slot></template></aside>"
[FAIL] ShadowRoot.getInnerHTML() on <aside>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<aside><template shadowroot=\\"closed\\"><slot></slot></template></aside>" but got "<aside><template shadowrootmode=\\"closed\\"><slot></slot></template></aside>"
[FAIL] ShadowRoot.getInnerHTML() on <aside>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<aside><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></aside>" but got "<aside><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></aside>"
[FAIL] ShadowRoot.getInnerHTML() on <aside>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<aside><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></aside>" but got "<aside><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></aside>"
[FAIL] Element.getInnerHTML() on <blockquote>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<blockquote><template shadowroot=\\"open\\"><slot></slot></template></blockquote>" but got "<blockquote><template shadowrootmode=\\"open\\"><slot></slot></template></blockquote>"
[FAIL] Element.getInnerHTML() on <blockquote>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<blockquote><template shadowroot=\\"closed\\"><slot></slot></template></blockquote>" but got "<blockquote><template shadowrootmode=\\"closed\\"><slot></slot></template></blockquote>"
[FAIL] Element.getInnerHTML() on <blockquote>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<blockquote><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></blockquote>" but got "<blockquote><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></blockquote>"
[FAIL] Element.getInnerHTML() on <blockquote>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<blockquote><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></blockquote>" but got "<blockquote><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></blockquote>"
[FAIL] ShadowRoot.getInnerHTML() on <blockquote>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<blockquote><template shadowroot=\\"open\\"><slot></slot></template></blockquote>" but got "<blockquote><template shadowrootmode=\\"open\\"><slot></slot></template></blockquote>"
[FAIL] ShadowRoot.getInnerHTML() on <blockquote>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<blockquote><template shadowroot=\\"closed\\"><slot></slot></template></blockquote>" but got "<blockquote><template shadowrootmode=\\"closed\\"><slot></slot></template></blockquote>"
[FAIL] ShadowRoot.getInnerHTML() on <blockquote>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<blockquote><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></blockquote>" but got "<blockquote><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></blockquote>"
[FAIL] ShadowRoot.getInnerHTML() on <blockquote>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<blockquote><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></blockquote>" but got "<blockquote><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></blockquote>"
[FAIL] Element.getInnerHTML() on <body>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<body><template shadowroot=\\"open\\"><slot></slot></template></body>" but got "<body><template shadowrootmode=\\"open\\"><slot></slot></template></body>"
[FAIL] Element.getInnerHTML() on <body>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<body><template shadowroot=\\"closed\\"><slot></slot></template></body>" but got "<body><template shadowrootmode=\\"closed\\"><slot></slot></template></body>"
[FAIL] Element.getInnerHTML() on <body>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<body><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></body>" but got "<body><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></body>"
[FAIL] Element.getInnerHTML() on <body>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<body><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></body>" but got "<body><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></body>"
[FAIL] ShadowRoot.getInnerHTML() on <body>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<body><template shadowroot=\\"open\\"><slot></slot></template></body>" but got "<body><template shadowrootmode=\\"open\\"><slot></slot></template></body>"
[FAIL] ShadowRoot.getInnerHTML() on <body>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<body><template shadowroot=\\"closed\\"><slot></slot></template></body>" but got "<body><template shadowrootmode=\\"closed\\"><slot></slot></template></body>"
[FAIL] ShadowRoot.getInnerHTML() on <body>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<body><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></body>" but got "<body><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></body>"
[FAIL] ShadowRoot.getInnerHTML() on <body>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<body><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></body>" but got "<body><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></body>"
[FAIL] Element.getInnerHTML() on <div>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<div><template shadowroot=\\"open\\"><slot></slot></template></div>" but got "<div><template shadowrootmode=\\"open\\"><slot></slot></template></div>"
[FAIL] Element.getInnerHTML() on <div>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<div><template shadowroot=\\"closed\\"><slot></slot></template></div>" but got "<div><template shadowrootmode=\\"closed\\"><slot></slot></template></div>"
[FAIL] Element.getInnerHTML() on <div>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<div><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></div>" but got "<div><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></div>"
[FAIL] Element.getInnerHTML() on <div>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<div><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></div>" but got "<div><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></div>"
[FAIL] ShadowRoot.getInnerHTML() on <div>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<div><template shadowroot=\\"open\\"><slot></slot></template></div>" but got "<div><template shadowrootmode=\\"open\\"><slot></slot></template></div>"
[FAIL] ShadowRoot.getInnerHTML() on <div>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<div><template shadowroot=\\"closed\\"><slot></slot></template></div>" but got "<div><template shadowrootmode=\\"closed\\"><slot></slot></template></div>"
[FAIL] ShadowRoot.getInnerHTML() on <div>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<div><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></div>" but got "<div><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></div>"
[FAIL] ShadowRoot.getInnerHTML() on <div>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<div><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></div>" but got "<div><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></div>"
[FAIL] Element.getInnerHTML() on <footer>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<footer><template shadowroot=\\"open\\"><slot></slot></template></footer>" but got "<footer><template shadowrootmode=\\"open\\"><slot></slot></template></footer>"
[FAIL] Element.getInnerHTML() on <footer>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<footer><template shadowroot=\\"closed\\"><slot></slot></template></footer>" but got "<footer><template shadowrootmode=\\"closed\\"><slot></slot></template></footer>"
[FAIL] Element.getInnerHTML() on <footer>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<footer><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></footer>" but got "<footer><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></footer>"
[FAIL] Element.getInnerHTML() on <footer>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<footer><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></footer>" but got "<footer><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></footer>"
[FAIL] ShadowRoot.getInnerHTML() on <footer>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<footer><template shadowroot=\\"open\\"><slot></slot></template></footer>" but got "<footer><template shadowrootmode=\\"open\\"><slot></slot></template></footer>"
[FAIL] ShadowRoot.getInnerHTML() on <footer>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<footer><template shadowroot=\\"closed\\"><slot></slot></template></footer>" but got "<footer><template shadowrootmode=\\"closed\\"><slot></slot></template></footer>"
[FAIL] ShadowRoot.getInnerHTML() on <footer>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<footer><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></footer>" but got "<footer><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></footer>"
[FAIL] ShadowRoot.getInnerHTML() on <footer>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<footer><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></footer>" but got "<footer><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></footer>"
[FAIL] Element.getInnerHTML() on <h1>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<h1><template shadowroot=\\"open\\"><slot></slot></template></h1>" but got "<h1><template shadowrootmode=\\"open\\"><slot></slot></template></h1>"
[FAIL] Element.getInnerHTML() on <h1>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<h1><template shadowroot=\\"closed\\"><slot></slot></template></h1>" but got "<h1><template shadowrootmode=\\"closed\\"><slot></slot></template></h1>"
[FAIL] Element.getInnerHTML() on <h1>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<h1><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h1>" but got "<h1><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h1>"
[FAIL] Element.getInnerHTML() on <h1>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<h1><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h1>" but got "<h1><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h1>"
[FAIL] ShadowRoot.getInnerHTML() on <h1>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<h1><template shadowroot=\\"open\\"><slot></slot></template></h1>" but got "<h1><template shadowrootmode=\\"open\\"><slot></slot></template></h1>"
[FAIL] ShadowRoot.getInnerHTML() on <h1>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<h1><template shadowroot=\\"closed\\"><slot></slot></template></h1>" but got "<h1><template shadowrootmode=\\"closed\\"><slot></slot></template></h1>"
[FAIL] ShadowRoot.getInnerHTML() on <h1>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<h1><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h1>" but got "<h1><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h1>"
[FAIL] ShadowRoot.getInnerHTML() on <h1>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<h1><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h1>" but got "<h1><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h1>"
[FAIL] Element.getInnerHTML() on <h2>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<h2><template shadowroot=\\"open\\"><slot></slot></template></h2>" but got "<h2><template shadowrootmode=\\"open\\"><slot></slot></template></h2>"
[FAIL] Element.getInnerHTML() on <h2>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<h2><template shadowroot=\\"closed\\"><slot></slot></template></h2>" but got "<h2><template shadowrootmode=\\"closed\\"><slot></slot></template></h2>"
[FAIL] Element.getInnerHTML() on <h2>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<h2><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h2>" but got "<h2><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h2>"
[FAIL] Element.getInnerHTML() on <h2>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<h2><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h2>" but got "<h2><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h2>"
[FAIL] ShadowRoot.getInnerHTML() on <h2>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<h2><template shadowroot=\\"open\\"><slot></slot></template></h2>" but got "<h2><template shadowrootmode=\\"open\\"><slot></slot></template></h2>"
[FAIL] ShadowRoot.getInnerHTML() on <h2>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<h2><template shadowroot=\\"closed\\"><slot></slot></template></h2>" but got "<h2><template shadowrootmode=\\"closed\\"><slot></slot></template></h2>"
[FAIL] ShadowRoot.getInnerHTML() on <h2>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<h2><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h2>" but got "<h2><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h2>"
[FAIL] ShadowRoot.getInnerHTML() on <h2>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<h2><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h2>" but got "<h2><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h2>"
[FAIL] Element.getInnerHTML() on <h3>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<h3><template shadowroot=\\"open\\"><slot></slot></template></h3>" but got "<h3><template shadowrootmode=\\"open\\"><slot></slot></template></h3>"
[FAIL] Element.getInnerHTML() on <h3>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<h3><template shadowroot=\\"closed\\"><slot></slot></template></h3>" but got "<h3><template shadowrootmode=\\"closed\\"><slot></slot></template></h3>"
[FAIL] Element.getInnerHTML() on <h3>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<h3><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h3>" but got "<h3><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h3>"
[FAIL] Element.getInnerHTML() on <h3>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<h3><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h3>" but got "<h3><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h3>"
[FAIL] ShadowRoot.getInnerHTML() on <h3>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<h3><template shadowroot=\\"open\\"><slot></slot></template></h3>" but got "<h3><template shadowrootmode=\\"open\\"><slot></slot></template></h3>"
[FAIL] ShadowRoot.getInnerHTML() on <h3>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<h3><template shadowroot=\\"closed\\"><slot></slot></template></h3>" but got "<h3><template shadowrootmode=\\"closed\\"><slot></slot></template></h3>"
[FAIL] ShadowRoot.getInnerHTML() on <h3>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<h3><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h3>" but got "<h3><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h3>"
[FAIL] ShadowRoot.getInnerHTML() on <h3>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<h3><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h3>" but got "<h3><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h3>"
[FAIL] Element.getInnerHTML() on <h4>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<h4><template shadowroot=\\"open\\"><slot></slot></template></h4>" but got "<h4><template shadowrootmode=\\"open\\"><slot></slot></template></h4>"
[FAIL] Element.getInnerHTML() on <h4>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<h4><template shadowroot=\\"closed\\"><slot></slot></template></h4>" but got "<h4><template shadowrootmode=\\"closed\\"><slot></slot></template></h4>"
[FAIL] Element.getInnerHTML() on <h4>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<h4><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h4>" but got "<h4><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h4>"
[FAIL] Element.getInnerHTML() on <h4>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<h4><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h4>" but got "<h4><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h4>"
[FAIL] ShadowRoot.getInnerHTML() on <h4>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<h4><template shadowroot=\\"open\\"><slot></slot></template></h4>" but got "<h4><template shadowrootmode=\\"open\\"><slot></slot></template></h4>"
[FAIL] ShadowRoot.getInnerHTML() on <h4>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<h4><template shadowroot=\\"closed\\"><slot></slot></template></h4>" but got "<h4><template shadowrootmode=\\"closed\\"><slot></slot></template></h4>"
[FAIL] ShadowRoot.getInnerHTML() on <h4>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<h4><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h4>" but got "<h4><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h4>"
[FAIL] ShadowRoot.getInnerHTML() on <h4>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<h4><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h4>" but got "<h4><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h4>"
[FAIL] Element.getInnerHTML() on <h5>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<h5><template shadowroot=\\"open\\"><slot></slot></template></h5>" but got "<h5><template shadowrootmode=\\"open\\"><slot></slot></template></h5>"
[FAIL] Element.getInnerHTML() on <h5>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<h5><template shadowroot=\\"closed\\"><slot></slot></template></h5>" but got "<h5><template shadowrootmode=\\"closed\\"><slot></slot></template></h5>"
[FAIL] Element.getInnerHTML() on <h5>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<h5><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h5>" but got "<h5><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h5>"
[FAIL] Element.getInnerHTML() on <h5>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<h5><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h5>" but got "<h5><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h5>"
[FAIL] ShadowRoot.getInnerHTML() on <h5>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<h5><template shadowroot=\\"open\\"><slot></slot></template></h5>" but got "<h5><template shadowrootmode=\\"open\\"><slot></slot></template></h5>"
[FAIL] ShadowRoot.getInnerHTML() on <h5>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<h5><template shadowroot=\\"closed\\"><slot></slot></template></h5>" but got "<h5><template shadowrootmode=\\"closed\\"><slot></slot></template></h5>"
[FAIL] ShadowRoot.getInnerHTML() on <h5>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<h5><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h5>" but got "<h5><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h5>"
[FAIL] ShadowRoot.getInnerHTML() on <h5>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<h5><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h5>" but got "<h5><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h5>"
[FAIL] Element.getInnerHTML() on <h6>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<h6><template shadowroot=\\"open\\"><slot></slot></template></h6>" but got "<h6><template shadowrootmode=\\"open\\"><slot></slot></template></h6>"
[FAIL] Element.getInnerHTML() on <h6>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<h6><template shadowroot=\\"closed\\"><slot></slot></template></h6>" but got "<h6><template shadowrootmode=\\"closed\\"><slot></slot></template></h6>"
[FAIL] Element.getInnerHTML() on <h6>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<h6><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h6>" but got "<h6><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h6>"
[FAIL] Element.getInnerHTML() on <h6>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<h6><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h6>" but got "<h6><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h6>"
[FAIL] ShadowRoot.getInnerHTML() on <h6>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<h6><template shadowroot=\\"open\\"><slot></slot></template></h6>" but got "<h6><template shadowrootmode=\\"open\\"><slot></slot></template></h6>"
[FAIL] ShadowRoot.getInnerHTML() on <h6>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<h6><template shadowroot=\\"closed\\"><slot></slot></template></h6>" but got "<h6><template shadowrootmode=\\"closed\\"><slot></slot></template></h6>"
[FAIL] ShadowRoot.getInnerHTML() on <h6>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<h6><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h6>" but got "<h6><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h6>"
[FAIL] ShadowRoot.getInnerHTML() on <h6>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<h6><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h6>" but got "<h6><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></h6>"
[FAIL] Element.getInnerHTML() on <header>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<header><template shadowroot=\\"open\\"><slot></slot></template></header>" but got "<header><template shadowrootmode=\\"open\\"><slot></slot></template></header>"
[FAIL] Element.getInnerHTML() on <header>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<header><template shadowroot=\\"closed\\"><slot></slot></template></header>" but got "<header><template shadowrootmode=\\"closed\\"><slot></slot></template></header>"
[FAIL] Element.getInnerHTML() on <header>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<header><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></header>" but got "<header><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></header>"
[FAIL] Element.getInnerHTML() on <header>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<header><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></header>" but got "<header><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></header>"
[FAIL] ShadowRoot.getInnerHTML() on <header>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<header><template shadowroot=\\"open\\"><slot></slot></template></header>" but got "<header><template shadowrootmode=\\"open\\"><slot></slot></template></header>"
[FAIL] ShadowRoot.getInnerHTML() on <header>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<header><template shadowroot=\\"closed\\"><slot></slot></template></header>" but got "<header><template shadowrootmode=\\"closed\\"><slot></slot></template></header>"
[FAIL] ShadowRoot.getInnerHTML() on <header>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<header><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></header>" but got "<header><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></header>"
[FAIL] ShadowRoot.getInnerHTML() on <header>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<header><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></header>" but got "<header><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></header>"
[FAIL] Element.getInnerHTML() on <main>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<main><template shadowroot=\\"open\\"><slot></slot></template></main>" but got "<main><template shadowrootmode=\\"open\\"><slot></slot></template></main>"
[FAIL] Element.getInnerHTML() on <main>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<main><template shadowroot=\\"closed\\"><slot></slot></template></main>" but got "<main><template shadowrootmode=\\"closed\\"><slot></slot></template></main>"
[FAIL] Element.getInnerHTML() on <main>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<main><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></main>" but got "<main><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></main>"
[FAIL] Element.getInnerHTML() on <main>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<main><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></main>" but got "<main><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></main>"
[FAIL] ShadowRoot.getInnerHTML() on <main>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<main><template shadowroot=\\"open\\"><slot></slot></template></main>" but got "<main><template shadowrootmode=\\"open\\"><slot></slot></template></main>"
[FAIL] ShadowRoot.getInnerHTML() on <main>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<main><template shadowroot=\\"closed\\"><slot></slot></template></main>" but got "<main><template shadowrootmode=\\"closed\\"><slot></slot></template></main>"
[FAIL] ShadowRoot.getInnerHTML() on <main>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<main><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></main>" but got "<main><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></main>"
[FAIL] ShadowRoot.getInnerHTML() on <main>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<main><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></main>" but got "<main><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></main>"
[FAIL] Element.getInnerHTML() on <nav>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<nav><template shadowroot=\\"open\\"><slot></slot></template></nav>" but got "<nav><template shadowrootmode=\\"open\\"><slot></slot></template></nav>"
[FAIL] Element.getInnerHTML() on <nav>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<nav><template shadowroot=\\"closed\\"><slot></slot></template></nav>" but got "<nav><template shadowrootmode=\\"closed\\"><slot></slot></template></nav>"
[FAIL] Element.getInnerHTML() on <nav>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<nav><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></nav>" but got "<nav><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></nav>"
[FAIL] Element.getInnerHTML() on <nav>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<nav><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></nav>" but got "<nav><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></nav>"
[FAIL] ShadowRoot.getInnerHTML() on <nav>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<nav><template shadowroot=\\"open\\"><slot></slot></template></nav>" but got "<nav><template shadowrootmode=\\"open\\"><slot></slot></template></nav>"
[FAIL] ShadowRoot.getInnerHTML() on <nav>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<nav><template shadowroot=\\"closed\\"><slot></slot></template></nav>" but got "<nav><template shadowrootmode=\\"closed\\"><slot></slot></template></nav>"
[FAIL] ShadowRoot.getInnerHTML() on <nav>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<nav><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></nav>" but got "<nav><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></nav>"
[FAIL] ShadowRoot.getInnerHTML() on <nav>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<nav><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></nav>" but got "<nav><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></nav>"
[FAIL] Element.getInnerHTML() on <p>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<p><template shadowroot=\\"open\\"><slot></slot></template></p>" but got "<p><template shadowrootmode=\\"open\\"><slot></slot></template></p>"
[FAIL] Element.getInnerHTML() on <p>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<p><template shadowroot=\\"closed\\"><slot></slot></template></p>" but got "<p><template shadowrootmode=\\"closed\\"><slot></slot></template></p>"
[FAIL] Element.getInnerHTML() on <p>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<p><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></p>" but got "<p><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></p>"
[FAIL] Element.getInnerHTML() on <p>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<p><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></p>" but got "<p><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></p>"
[FAIL] ShadowRoot.getInnerHTML() on <p>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<p><template shadowroot=\\"open\\"><slot></slot></template></p>" but got "<p><template shadowrootmode=\\"open\\"><slot></slot></template></p>"
[FAIL] ShadowRoot.getInnerHTML() on <p>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<p><template shadowroot=\\"closed\\"><slot></slot></template></p>" but got "<p><template shadowrootmode=\\"closed\\"><slot></slot></template></p>"
[FAIL] ShadowRoot.getInnerHTML() on <p>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<p><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></p>" but got "<p><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></p>"
[FAIL] ShadowRoot.getInnerHTML() on <p>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<p><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></p>" but got "<p><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></p>"
[FAIL] Element.getInnerHTML() on <section>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<section><template shadowroot=\\"open\\"><slot></slot></template></section>" but got "<section><template shadowrootmode=\\"open\\"><slot></slot></template></section>"
[FAIL] Element.getInnerHTML() on <section>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<section><template shadowroot=\\"closed\\"><slot></slot></template></section>" but got "<section><template shadowrootmode=\\"closed\\"><slot></slot></template></section>"
[FAIL] Element.getInnerHTML() on <section>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<section><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></section>" but got "<section><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></section>"
[FAIL] Element.getInnerHTML() on <section>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<section><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></section>" but got "<section><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></section>"
[FAIL] ShadowRoot.getInnerHTML() on <section>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<section><template shadowroot=\\"open\\"><slot></slot></template></section>" but got "<section><template shadowrootmode=\\"open\\"><slot></slot></template></section>"
[FAIL] ShadowRoot.getInnerHTML() on <section>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<section><template shadowroot=\\"closed\\"><slot></slot></template></section>" but got "<section><template shadowrootmode=\\"closed\\"><slot></slot></template></section>"
[FAIL] ShadowRoot.getInnerHTML() on <section>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<section><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></section>" but got "<section><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></section>"
[FAIL] ShadowRoot.getInnerHTML() on <section>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<section><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></section>" but got "<section><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></section>"
[FAIL] Element.getInnerHTML() on <span>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<span><template shadowroot=\\"open\\"><slot></slot></template></span>" but got "<span><template shadowrootmode=\\"open\\"><slot></slot></template></span>"
[FAIL] Element.getInnerHTML() on <span>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<span><template shadowroot=\\"closed\\"><slot></slot></template></span>" but got "<span><template shadowrootmode=\\"closed\\"><slot></slot></template></span>"
[FAIL] Element.getInnerHTML() on <span>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<span><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></span>" but got "<span><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></span>"
[FAIL] Element.getInnerHTML() on <span>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<span><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></span>" but got "<span><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></span>"
[FAIL] ShadowRoot.getInnerHTML() on <span>, with mode=open, delegatesFocus=false.
  assert_equals: The default for includeShadowRoots should be true expected "<span><template shadowroot=\\"open\\"><slot></slot></template></span>" but got "<span><template shadowrootmode=\\"open\\"><slot></slot></template></span>"
[FAIL] ShadowRoot.getInnerHTML() on <span>, with mode=closed, delegatesFocus=false.
  assert_equals: expected "<span><template shadowroot=\\"closed\\"><slot></slot></template></span>" but got "<span><template shadowrootmode=\\"closed\\"><slot></slot></template></span>"
[FAIL] ShadowRoot.getInnerHTML() on <span>, with mode=open, delegatesFocus=true.
  assert_equals: The default for includeShadowRoots should be true expected "<span><template shadowroot=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></span>" but got "<span><template shadowrootmode=\\"open\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></span>"
[FAIL] ShadowRoot.getInnerHTML() on <span>, with mode=closed, delegatesFocus=true.
  assert_equals: expected "<span><template shadowroot=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></span>" but got "<span><template shadowrootmode=\\"closed\\" shadowrootdelegatesfocus=\\"\\"><slot></slot></template></span>"
Harness: the test ran to completion.
